backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
dj-database-url==2.1.0
django-cors-headers==4.7.0
django-filter==25.1
django-simple-history==3.8.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
gunicorn==21.2.0
openpyxl==3.1.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.6.0
